kmsan: support compat syscalls
compat syscalls allocate parameters for real syscalls on the kernel
stack, cast them to __user pointers and pass those to the real syscalls.
As a result, those parameters are initialized with copy_to_user(), which
in this case copies data between kernel structs.
To avoid false positives, we special-case copy_to_user() to kernel
addresses, treating it as a memcpy().